refactoring util + bench not compiling
