+++ pallets/thea/Cargo.toml
@@ -14,20 +14,25 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 parity-scale-codec = { workspace = true, default-features = false, features = ["max-encoded-len"] }
 scale-info = { workspace = true, default-features = false, features = ["derive"] }
-
 frame-support = { workspace = true, default-features = false }
 frame-system = { workspace = true, default-features = false }
-frame-benchmarking = { workspace = true, default-features = false }
-
 sp-core = { workspace = true, default-features = false }
 sp-std = { workspace = true, default-features = false }
 sp-runtime = { workspace = true, default-features = false }
-
 asset-handler = {path = "../asset-handler", default-features = false}
 thea-primitives = { path = "../../thea-primitives", default-features = false }
 thea-staking = { path = "../thea-staking", default-features = false }
 xcm = { workspace = true, default-features = false }
-pallet-balances = { default-features = false, workspace = true } # TODO: optional for PR, optional = true }
+blst =  { verisou = "0.3.10", optional = true }
+sp-io = { workspace = true, default-features = false, optional = true }
+sp-keystore = { workspace = true, default-features = false, optional = true }
+pallet-assets = { workspace = true, default-features = false, optional = true }
+pallet-staking = { default-features = false, workspace = true, optional = true }
+pallet-balances = { default-features = false, workspace = true , optional = true }
+chainbridge = { path = "../chainbridge", default-features = false, optional = true }
+frame-benchmarking = { workspace = true, default-features = false, optional = true }
+sp-application-crypto = { workspace = true, default-features = false, optional = true }
+pallet-staking-reward-curve = { workspace = true, default-features = false, optional = true }
 
 [dev-dependencies]
 pallet-staking-reward-curve = { workspace = true }
@@ -56,12 +61,20 @@ std = [
     "xcm/std",
     "thea-staking/std"
 ]
-
 runtime-benchmarks = [
+    "frame-benchmarking/runtime-benchmarks",
+    "pallet-balances/runtime-benchmarks",
+    "asset-handler/runtime-benchmarks",
+    "pallet-assets/runtime-benchmarks",
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
-    "frame-benchmarking/runtime-benchmarks",
     "sp-runtime/runtime-benchmarks",
-    "asset-handler/runtime-benchmarks",
-    "pallet-balances/runtime-benchmarks",
+    "pallet-staking-reward-curve",
+    "sp-application-crypto",
+    "pallet-staking",
+    "chainbridge",
+    "sp-keystore",
+    "sp-io",
+    "blst"
 ]
+try-runtime = ["frame-support/try-runtime"]
